What's done Выполнен дебаг функционала получения сводной ведомости.

+++ src/main/java/net/thumbtack/onlineshop/dto/Request/HistoryListDtoRequest.java
@@ -28,7 +28,7 @@ public class HistoryListDtoRequest {
         return offset;
     }
 
-    public void setOffset(int offset) {
+    public void setOffset(long offset) {
         this.offset = offset;
     }
